update long to int

+++ src/main/java/com/ngo/fundraiser/entity/Theme.java
@@ -10,7 +10,7 @@ public class Theme {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long themeID;
+    private int themeID;
     private String campaignName;
     private String keywords;
 }
